# docs/swagger.yaml
basePath: /
definitions:
  crud.CartResponse:
    properties:
      count:
        description: ← теперь общее количество товаров!
        type: integer
      items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      total:
        type: integer
    type: object
...
  models.CartItem:
    properties:
      availability:
        type: string
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: string
      description:
        type: string
      features:
        items:
          type: string
        type: array
      id:
        type: string
      image:
        items:
          type: string
        type: array
      manufacturer:
        $ref: '#/definitions/models.Manufacturer'
      manufacturerId:
        type: string
      name:
        type: string
      oldPrice:
        type: integer
      price:
        type: integer
      quantity:
        type: integer
      rating:
        type: number
      reviews:
        type: integer
      sku:
        type: string
      slug:
        type: string
      stock:
        type: integer
    type: object
  models.Category:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      parentId:
        type: string
      slug:
        type: string
    type: object
...
  models.Manufacturer:
    properties:
      id:
        type: string
      logo:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Noble Group Services e-commerce platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Noble Group Services API
  version: "1.0"
paths:
  /cart:
    delete:
      description: Remove all items from the cart
      parameters:
      - description: Session ID
        in: header
        name: X-Session-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/crud.CartResponse'
      summary: Clear cart
      tags:
      - cart
    get:
      description: Get the current session's cart
      parameters:
      - description: Session ID
        in: header
        name: X-Session-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/crud.CartResponse'
      summary: Get current cart
      tags:
      - cart
    post:
      consumes:
      - application/json
      description: Add a product to the cart
      parameters:
      - description: Session ID
        in: header
        name: X-Session-ID
        type: string
      - description: Product ID and Quantity
        in: body
        name: request
        required: true
        schema:
          properties:
            productId:
              type: string
            quantity:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/crud.CartResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
      summary: Add item to cart
      tags:
      - cart
  /cart/{id}:
    delete:
      description: Remove a product from the cart
      parameters:
      - description: Session ID
        in: header
        name: X-Session-ID
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/crud.CartResponse'
        "404":
          description: Item not found
          schema:
            type: string
      summary: Remove item from cart
      tags:
      - cart
    patch:
      consumes:
      - application/json
      description: Update the quantity of a product in the cart
      parameters:
      - description: Session ID
        in: header
        name: X-Session-ID
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: New Quantity
        in: body
        name: request
        required: true
        schema:
          properties:
            quantity:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/crud.CartResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: Update cart item quantity
      tags:
      - cart
...
  /products/categories:
    get:
      description: Get a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create a category
      tags:
      - categories
  /products/categories/{id}:
    delete:
      description: Delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Category not found
          schema:
            type: string
      summary: Delete category
      tags:
      - categories
    get:
      description: Get details of a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "404":
          description: Category not found
          schema:
            type: string
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Category not found
          schema:
            type: string
      summary: Update category
      tags:
      - categories
...
swagger: "2.0"